Refactor the wallet connect
