+++ apps/mobile/src/components/Login/ConnectWalletButton.tsx
@@ -1,5 +1,5 @@
 import { useNavigation } from '@react-navigation/native';
-import { useWalletConnectModal, WalletConnectModal } from '@walletconnect/modal-react-native';
+import { useWalletConnectModal } from '@walletconnect/modal-react-native';
 import { ethers } from 'ethers';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useLogin } from 'src/hooks/useLogin';
@@ -14,24 +14,9 @@ import { Button } from '../../components/Button';
 import { GalleryBottomSheetModalType } from '../GalleryBottomSheet/GalleryBottomSheetModal';
 import { WalletSelectorBottomSheet } from './WalletSelectorBottomSheet';
 
-const projectId = '93ea9c14da3ab92ac4b72d97c124b96c';
-
-const providerMetadata = {
-  name: 'Gallery',
-  description: 'Gallery Mobile App',
-  url: 'https://gallery.so',
-  icons: [
-    'https://f4shnljo4g7olt4wifnpdfz6752po37hr3waoaxakgpxqw64jojq.arweave.net/LyR2rS7hvuXPlkFa8Zc-_3T3b-eO7AcC4FGfeFvcS5M',
-  ],
-  redirect: {
-    native: 'applinks:gallery.so/',
-    universal: 'https://gallery.so',
-  },
-};
-
 export function ConnectWalletButton() {
   const navigation = useNavigation<LoginStackNavigatorProp>();
-  const { open, isConnected, provider, address } = useWalletConnectModal();
+  const { address, isConnected, provider } = useWalletConnectModal();
   const createNonce = useCreateNonce();
   const { pushToast } = useToastActions();
 
@@ -106,13 +91,6 @@ export function ConnectWalletButton() {
     }
   }, [address, createNonce, login, provider, pushToast, web3Provider, navigation, track]);
 
-  const handleEthereumConnect = useCallback(async () => {
-    if (isConnected) {
-      return provider?.disconnect();
-    }
-    open();
-  }, [isConnected, open, provider]);
-
   const handleButtonPress = useCallback(() => {
     bottomSheet.current?.present();
   }, []);
@@ -146,10 +124,8 @@ export function ConnectWalletButton() {
       <WalletSelectorBottomSheet
         ref={bottomSheet}
         isSignedIn={isSigningIn}
-        onConnectWallet={handleEthereumConnect}
         onDismiss={handleDismiss}
       />
-      <WalletConnectModal projectId={projectId} providerMetadata={providerMetadata} />
     </>
   );
 }
